feat: move file to domain directory
