+++ backend/src/main/java/com/cocktailmasters/backend/report/controller/ReportController.java
@@ -1,0 +1,5 @@
+package com.cocktailmasters.backend.report.controller;
+
+public class ReportController {
+
+}
